// lib/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const SkillCategory = {
  HARD: 'HARD',
  SOFT: 'SOFT',
  META: 'META'
} as const

export type SkillCategory = (typeof SkillCategory)[keyof typeof SkillCategory]


export const AssessmentStatus = {
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  REVIEW_PENDING: 'REVIEW_PENDING'
} as const

export type AssessmentStatus = (typeof AssessmentStatus)[keyof typeof AssessmentStatus]